requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
et-xmlfile==1.1.0
Flask==2.0.2
Flask-Cors==3.0.10
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
newsapi-python==0.2.6
numpy==1.21.3
openai==0.10.5
openpyxl==3.0.9
pandas==1.3.4
pandas-stubs==1.2.0.35
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
six==1.16.0
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
